linebot_test_version/requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
alabaster==0.7.12
altgraph==0.17
anaconda-client==1.7.2
anaconda-navigator==1.9.7
anaconda-project==0.8.3
appnope==0.1.0
appscript==1.0.1
asgiref==3.4.1
asn1crypto==1.0.1
astroid==2.3.1
astropy==3.2.2
async-timeout==4.0.1
asynctest==0.13.0
atomicwrites==1.3.0
attrs==19.2.0
Babel==2.7.0
backcall==0.1.0
backports.functools-lru-cache==1.5
backports.os==0.1.1
backports.shutil-get-terminal-size==1.0.0
backports.tempfile==1.0
backports.weakref==1.0.post1
beautifulsoup4==4.8.0
bitarray==1.0.1
bkcharts==0.2
bleach==3.1.0
bokeh==1.3.4
boto==2.49.0
Bottleneck==1.2.1
branca==0.4.2
certifi==2019.9.11
cffi==1.12.3
chardet==3.0.4
charset-normalizer==2.0.7
Click==7.0
click-plugins==1.1.1
cligj==0.7.2
cloudpickle==1.2.2
clyent==1.2.2
cmdstanpy==0.9.5
colorama==0.4.1
conda==4.9.2
conda-build==3.18.9
conda-package-handling==1.6.0
conda-verify==3.4.2
contextlib2==0.6.0
convertdate @ file:///home/conda/feedstock_root/build_artifacts/convertdate_1605102623033/work
cryptography==2.7
cycler==0.10.0
Cython==0.29.13
cytoolz==0.10.0
dask==2.5.2
decorator==4.4.0
defusedxml==0.6.0
distributed==2.5.2
Django==3.2.9
docker==4.3.1
docopt==0.6.2
docutils==0.15.2
entrypoints==0.3
ephem==3.7.7.1
et-xmlfile==1.0.1
fastcache==1.1.0
fbprophet==0.7.1
filelock==3.0.12
Fiona==1.8.20
Flask==1.1.1
folium==0.12.1
frozenlist==1.2.0
fsspec==0.5.2
future==0.17.1
gensim==4.1.2
geopandas==0.9.0
gevent==1.4.0
glob2==0.7
gmpy2==2.0.8
graphviz==0.14
greenlet==0.4.15
gunicorn==20.1.0
h5py==2.9.0
HeapDict==1.0.1
holidays @ file:///home/conda/feedstock_root/build_artifacts/holidays_1595448845196/work
html5lib==1.0.1
idna==2.8
imageio==2.6.0
imagesize==1.1.0
importlib-metadata==0.23
inflection==0.5.1
ipykernel==5.1.2
ipython==7.8.0
ipython_genutils==0.2.0
ipywidgets==7.5.1
isort==4.3.21
itsdangerous==1.1.0
jdcal==1.4.1
jedi==0.15.1
jieba==0.42.1
Jinja2==2.10.3
joblib==1.0.1
json5==0.8.5
jsonschema==3.0.2
jupyter==1.0.0
jupyter-client==5.3.3
jupyter-console==6.0.0
jupyter-core==4.5.0
jupyterlab==1.1.4
jupyterlab-server==1.0.6
keyring==18.0.0
kiwisolver==1.3.2
korean-lunar-calendar @ file:///home/conda/feedstock_root/build_artifacts/korean_lunar_calendar_1589354365811/work
lazy-object-proxy==1.4.2
libarchive-c==2.8
lief==0.9.0
line-bot-sdk==2.0.1
llvmlite==0.29.0
locket==0.2.0
LunarCalendar==0.0.9
lunardate==0.2.0
lxml==4.4.1
macholib @ file:///tmp/build/80754af9/macholib_1597435072973/work
MarkupSafe==1.1.1
matplotlib==3.4.3
mccabe==0.6.1
mistune==0.8.4
mkl-fft==1.0.14
mkl-random==1.1.0
mkl-service==2.3.0
mlxtend==0.19.0
mock==3.0.5
more-itertools==7.2.0
mpmath==1.1.0
msgpack==0.6.1
multidict==5.2.0
multipledispatch==0.6.0
multitasking==0.0.9
munch==2.5.0
navigator-updater==0.2.1
nbconvert==5.6.0
nbformat==4.4.0
networkx==2.3
nltk==3.4.5
nose==1.3.7
notebook==6.0.1
numba==0.45.1
numexpr==2.7.0
numpy==1.21.2
numpydoc==0.9.1
olefile==0.46
openpyxl==3.0.0
packaging==19.2
pandas==1.2.1
pandocfilters==1.4.2
parso==0.5.1
partd==1.0.0
path.py==12.0.1
pathlib2==2.3.5
patsy==0.5.1
pep8==1.7.1
pexpect==4.7.0
pickleshare==0.7.5
Pillow==8.3.2
pipreqs==0.4.10
pkginfo==1.5.0.1
plotly==4.14.3
pluggy==0.13.0
ply==3.11
prometheus-client==0.7.1
prompt-toolkit==2.0.10
psutil==5.6.3
ptyprocess==0.6.0
py==1.8.0
pycodestyle==2.5.0
pycosat==0.6.3
pycparser==2.19
pycrypto==2.6.1
pycurl==7.43.0.3
pydot==1.4.1
pydotplus @ file:///tmp/build/80754af9/pydotplus_1594390100614/work
pyflakes==2.1.1
pygame==1.9.6
Pygments==2.4.2
pygraphviz==1.3
pyinstaller==4.0
pyinstaller-hooks-contrib==2020.9
pylint==2.4.2
PyMeeus @ file:///home/conda/feedstock_root/build_artifacts/pymeeus_1589222711601/work
pyodbc==4.0.27
pyOpenSSL==19.0.0
pyparsing==2.4.7
pyproj==3.1.0
pyrsistent==0.15.4
PySocks==1.7.1
pystan==2.19.1.1
pytest==5.2.1
pytest-arraydiff==0.3
pytest-astropy==0.5.0
pytest-doctestplus==0.4.0
pytest-openfiles==0.4.0
pytest-remotedata==0.3.2
python-dateutil==2.8.2
pytrends==4.7.3
pytz==2019.3
pywifi==1.1.12
PyYAML==5.1.2
pyzmq==18.1.0
QtAwesome==0.6.0
qtconsole==4.5.5
QtPy==1.9.0
Quandl==3.6.0
requests==2.22.0
retrying==1.3.3
rope==0.14.0
ruamel_yaml==0.15.46
scikit-image==0.15.0
scikit-learn==1.0
scipy==1.7.1
seaborn==0.9.0
Send2Trash==1.5.0
setuptools-git==1.2
Shapely==1.7.1
simplegeneric==0.8.1
singledispatch==3.4.0.3
six==1.16.0
sklearn==0.0
smart-open==5.2.1
snowballstemmer==2.0.0
sortedcollections==1.1.2
sortedcontainers==2.1.0
soupsieve==1.9.3
Sphinx==2.2.0
sphinxcontrib-applehelp==1.0.1
sphinxcontrib-devhelp==1.0.1
sphinxcontrib-htmlhelp==1.0.2
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.2
sphinxcontrib-serializinghtml==1.1.3
sphinxcontrib-websupport==1.1.2
spyder==3.3.6
spyder-kernels==0.5.2
SQLAlchemy==1.3.9
sqlparse==0.4.2
statsmodels==0.10.1
sympy==1.4
tables==3.5.2
tblib==1.4.0
terminado==0.8.2
testpath==0.4.2
textrank4zh==0.3
threadpoolctl==2.2.0
toolz==0.10.0
tornado==6.0.3
tqdm==4.36.1
traitlets==4.3.3
typing_extensions==4.0.0
unicodecsv==0.14.1
urllib3==1.24.2
wcwidth==0.1.7
webencodings==0.5.1
websocket-client==0.57.0
Werkzeug==0.16.0
widgetsnbextension==3.5.1
wordcloud==1.8.1
wrapt==1.11.2
wurlitzer==1.0.3
xlrd==1.2.0
XlsxWriter==1.2.1
xlwings==0.15.10
xlwt==1.3.0
yarg==0.1.9
yarl==1.7.2
yfinance==0.1.55
zict==1.0.0
zipp==0.6.0
